restructure and improve error messages
