+++ rascal2/widgets/__init__.py
@@ -1,3 +1,4 @@
 from rascal2.widgets.controls.view import ControlsWidget
+from rascal2.widgets.inputs import ValidatedInputWidget
 
-__all__ = ["ControlsWidget"]
+__all__ = ["ControlsWidget", "ValidatedInputWidget"]
