+++ rascal2/widgets/controls/view.py
@@ -1,13 +1,12 @@
 """Widget for setting up the Controls class."""
 
-from enum import Enum
 from typing import Callable
 
-from pydantic.fields import FieldInfo
 from PyQt6 import QtCore, QtGui, QtWidgets
 from RATapi.utils.enums import Procedures
 
 from rascal2.config import path_for
+from rascal2.widgets.inputs import ValidatedInputWidget
 from rascal2.widgets.controls.model import FitSettingsModel
 
 
@@ -50,8 +49,8 @@ class ControlsWidget(QtWidgets.QWidget):
         self.validation_label = QtWidgets.QLabel("")
         self.validation_label.setStyleSheet("color : red")
         self.validation_label.font().setPointSize(10)
-        self.validation_label.setMaximumHeight(50)
         self.validation_label.setWordWrap(True)
+        self.validation_label.setSizePolicy(QtWidgets.QSizePolicy.Policy.Fixed, QtWidgets.QSizePolicy.Policy.Maximum)
 
         # create box containing chi-squared value
         chi_layout = QtWidgets.QHBoxLayout()
@@ -82,10 +81,10 @@ class ControlsWidget(QtWidgets.QWidget):
         buttons_layout.addWidget(self.stop_button)
         buttons_layout.addLayout(procedure_layout)
         buttons_layout.addWidget(self.fit_settings_button)
-        buttons_layout.addWidget(self.validation_label)
         buttons_layout.setSpacing(20)
         procedure_box.addLayout(chi_layout)
         procedure_box.addLayout(buttons_layout)
+        procedure_box.addWidget(self.validation_label)
 
         widget_layout = QtWidgets.QHBoxLayout()
         widget_layout.addLayout(procedure_box)
@@ -121,9 +120,13 @@ class ControlsWidget(QtWidgets.QWidget):
         if toggled:
             invalid_inputs = self.fit_settings_layout.currentWidget().get_invalid_inputs()
             if invalid_inputs:
-                self.validation_label.setText(f"Invalid inputs:\n  {"\n  ".join(invalid_inputs)}")
+                self.validation_label.setText("Could not run due to invalid inputs:\n  "
+                                                f"{"\n  ".join(invalid_inputs)}\n"
+                                                "Fix these inputs and try again.\n"
+                                                "See fit settings for more details.\n")
                 self.run_button.toggle()
                 return
+
             self.fit_settings.setEnabled(False)
             self.procedure_dropdown.setEnabled(False)
             self.run_button.setEnabled(False)
@@ -138,85 +141,21 @@ class ControlsWidget(QtWidgets.QWidget):
             self.run_button.setEnabled(True)
             self.stop_button.setEnabled(False)
 
+    def clear_validation_label(self):
+        self.validation_label.setText("")
+
     def set_procedure(self, index: int):
         """Change the Controls procedure and update the table.
 
         Parameters
         ----------
         index : int
-            The index of the procedure to change to.
+            The index of the procedure to change to in the procedure list.
 
         """
         self.fit_settings_layout.setCurrentIndex(index)
         procedure = [p.value for p in Procedures][index]
         self.fit_settings_model.setData("procedure", procedure)
-
-
-class ValidatedInputWidget(QtWidgets.QWidget):
-    """Number input for Pydantic field with validation."""
-
-    def __init__(self, field_info: FieldInfo, parent=None):
-        super().__init__(parent=parent)
-        layout = QtWidgets.QVBoxLayout()
-        self.input_is_valid = True
-
-        # editor_data and change_editor_data are set to the getter and setter
-        # methods for the actual editor inside the widget
-        self.editor_data: Callable
-        self.change_editor_data: Callable
-        self.edited_signal: QtCore.pyqtSignal
-
-        # widget, getter, setter and change signal for different datatypes
-        editor_types = {
-            int: (QtWidgets.QSpinBox, "value", "setValue", "valueChanged"),
-            float: (QtWidgets.QDoubleSpinBox, "value", "setValue", "valueChanged"),
-            bool: (QtWidgets.QCheckBox, "isChecked", "setChecked", "checkStateChanged"),
-        }
-        defaults = (QtWidgets.QLineEdit, "text", "setText", "textChanged")
-
-        if issubclass(field_info.annotation, Enum):
-            self.editor = QtWidgets.QComboBox(self)
-            self.editor.addItems(str(e.value) for e in field_info.annotation)
-            self.editor_data = self.editor.currentText
-            self.change_editor_data = self.editor.setCurrentText
-            self.edited_signal = self.editor.currentTextChanged
-        else:
-            editor, getter, setter, signal = editor_types.get(field_info.annotation, defaults)
-            self.editor = editor(self)
-            self.editor_data = getattr(self.editor, getter)
-            self.change_editor_data = getattr(self.editor, setter)
-            self.edited_signal = getattr(self.editor, signal)
-        if isinstance(self.editor, (QtWidgets.QSpinBox, QtWidgets.QDoubleSpinBox)):
-            set_constraints(self.editor, field_info)
-
-        layout.addWidget(self.editor)
-
-        self.validation_box = QtWidgets.QLabel()
-        self.validation_box.setStyleSheet("QLabel { color : red; }")
-        self.validation_box.font().setPointSize(10)
-        self.validation_box.setWordWrap(True)
-        layout.addWidget(self.validation_box)
-
-        layout.setContentsMargins(5, 0, 0, 0)
-
-        self.setLayout(layout)
-        self.setSizePolicy(QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Fixed)
-
-    def set_validation_text(self, msg: str):
-        """Write a message in the validation box.
-
-        Parameters
-        ----------
-        msg: str
-            The message to write in the validation box.
-        """
-        self.validation_box.setText(msg)
-        if msg != "":
-            self.editor.setStyleSheet("color : red")
-            self.input_is_valid = False
-        else:
-            self.editor.setStyleSheet("")
-            self.input_is_valid = True
 
 
 class FitSettingsWidget(QtWidgets.QScrollArea):
@@ -260,7 +199,7 @@ class FitSettingsWidget(QtWidgets.QScrollArea):
 
         Parameters
         ----------
-        s: str
+        setting: str
             The setting to which the setter connects.
 
         """
@@ -285,18 +224,3 @@ class FitSettingsWidget(QtWidgets.QScrollArea):
 
         """
         return [s for s in self.rows if not self.rows[s].input_is_valid]
-
-
-def set_constraints(widget: QtWidgets.QSpinBox | QtWidgets.QDoubleSpinBox, field_info) -> None:
-    metadata = field_info.metadata
-    if isinstance(widget, QtWidgets.QDoubleSpinBox):
-        widget.setSingleStep(0.1)
-    for item in metadata:
-        # using a 'guessing attributes' method rather than importing the annotation classes
-        # which would add a dependency since they're from the annotated_types package
-        for attr in ["ge", "gt"]:
-            if hasattr(item, attr):
-                widget.setMinimum(getattr(item, attr))
-        for attr in ["le", "lt"]:
-            if hasattr(item, attr):
-                widget.setMaximum(getattr(item, attr))
